+++ src-inst/flow_storm/core.cljs
@@ -27,6 +27,7 @@
       :instrument-forms     [:cmd-err "[WARNING] :instrument-forms isn't supported in ClojureScript yet"]
       :re-run-flow          [:cmd-err "[WARNING] :re-run-flow isn't supported in ClojureScript yet"]
       :def-remote-value     [:cmd-ret [comm-id (def-remote-value-command args-map)]]
-      :get-remote-value     [:cmd-ret [comm-id (get-remote-value-command args-map)]] )
+      :get-remote-value     [:cmd-ret [comm-id (get-remote-value-command args-map)]]
+      :get-all-namespaces   [:cmd-err "[WARNING] :get-all-namespaces isn't supported in ClojureScript yet"])
 
     (catch js/Error e (js/console.error (utils/format "Error running command %s %s" method args-map) e))))