+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## New Features
 	
 	* Add def value button on every value panel to define the value so you can work with it at the repl
+	* Add namespaces browser with instrumentation capabilities
 	
 ### Changes