+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ## New Features
 	
+### Changes
+    
+### Bugs fixed
+
+## 2.2.68 (10-06-2022)
+
+## New Features
+	
 	* Add def value button on every value panel to define the value so you can work with it at the repl
 	* Add namespaces browser with instrumentation capabilities